Fixed code to get NaN for not natural numbers input

+++ Program.cs
@@ -10,11 +10,11 @@ var app = builder.Build();
 app.UseHealthChecks("/health");
 
 
-app.MapGet("/app/nur_matova_jan_gmail_com", (int x, int y) =>
+app.MapGet("/app/nur_matova_jan_gmail_com", (double x, double y) =>
 {
     if (x > 0 && y > 0 && x % 1 == 0 && y % 1 == 0)
     {
-        int lcm = 0, i = 1;
+        double lcm = 0, i = 1;
         while (lcm == 0)
         {
             if (x * i % y == 0)
